requirements.txt
# Core ML/DL libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.14.0
accelerate>=0.26.0

# Vector search and similarity
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
umap-learn>=0.5.3

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
requests>=2.31.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
rank-bm25>=0.2.2

# Jupyter support
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0

